Rename adapter Factory
