+++ library/Zend/Paginator/Adapter/Factory/DbSelectFactory.php
@@ -7,7 +7,7 @@
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
-namespace Zend\Paginator\Adapter\Service;
+namespace Zend\Paginator\Adapter\Factory;
 
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
